New version with on/off switching of LEDs
In the previous version the LED was turned on by placing the RFID chip close to the reader and lit for 5 seconds. Now the RFID chip can be used to turn the LEDs on and off. (Hopefully as now testing has been done to test the pull/push stuff from github)
So testing has to be done!